Set canned_acl on manifest files
Setting the canned_acl on the data files is nice but if it's not on the
manifest file then it doesn't do much good unless the same user is
writing and reading both. In some situations we have an individual IAM
user writing to S3 but using a shared role to read during the COPY
(using set_aws_role). This makes sure that the manifests are always
readable across accounts.